# Conflicts:
#	client/src/app/app-routing.module.ts
#	client/src/app/app.component.html
#	client/src/app/file-browser/components/file-browser.component.html
#	client/src/app/file-browser/components/file-view.component.ts
#	client/src/app/shared/services/error-handler.service.ts
#	client/src/app/shared/shared.module.ts

+++ appserver/app.py
@@ -31,9 +31,13 @@ def seed():
     con = db.engine.connect()
     trans = con.begin()
     for table in meta.sorted_tables:
-        con.execute(f'ALTER TABLE "{table.name}" DISABLE TRIGGER ALL;')
-        con.execute(f'TRUNCATE TABLE "{table.name}" RESTART IDENTITY CASCADE;')
-        con.execute(f'ALTER TABLE "{table.name}" ENABLE TRIGGER ALL;')
+        # don't want to truncate annotation stop words table
+        # because data was inserted in migration
+        # since it's needed on server when deployed
+        if table.name != 'annotation_stop_words' and table.name != 'alembic_version':
+            con.execute(f'ALTER TABLE "{table.name}" DISABLE TRIGGER ALL;')
+            con.execute(f'TRUNCATE TABLE "{table.name}" RESTART IDENTITY CASCADE;')
+            con.execute(f'ALTER TABLE "{table.name}" ENABLE TRIGGER ALL;')
     trans.commit()
 
     def find_existing_row(model, value):
@@ -47,6 +51,7 @@ def seed():
 
     with open("fixtures/seed.json", "r") as f:
         fixtures = json.load(f)
+
         for fixture in fixtures:
             module_name, class_name = fixture['model'].rsplit('.', 1)
             module = importlib.import_module(module_name)
